\name{lmridge-package}
\alias{lmridge-package}

\docType{package}

\title{Linear Ridge Regression}

\description{R package for fitting linear ridge regression models.}

\details{
This package contains functions for fitting linear ridge regression models, including functions for computation of different ridge related statistics (such as MSE, Var-Cov matrix, effective degrees of freedom and condition numbers), estimation of biasing parameter from different researchers, testing of ridge coefficients, model selection criteria, residuals, predicted values and fitted values. The package also includes function for plotting of ridge coefficients and different ridge statistics for selection of optimal value of biasing parameters \eqn{K}{K}.

For a complete list of functions, use \code{library(help="lmridge")}.
}

\author{Muhammad Imdadullah, Dr. Muhammad Aslam}

\keyword{ package }
